requirements.txt
pyserial
supabase
httpx
python-dotenv
